backend/requirements.txt
flask
sqlalchemy
pyjwt
python-dotenv